Provide *_result variants of reading functions with informative errors
Currently Ezjsonm does not provide users with the ability to properly
report errors. In particular, the error location inside the JSON
source is provided by Jsonm and thrown away by Ezjsonm.

This PR provides *_result variants of reading functions
that (instead of failing in badly-documented ways) return an
informative error type in the case of reading failure, along
with simple functions to provide basic error messages to users.

Note: this PR uses the "result" type from the standard library, and as
a consequence it bumps the required OCaml version from 4.04 to
4.08. If this is judged too strong, we could use a compatibility
"result" module instead.